Warnings smaller (#9910)
* reducing warnings

* reducing warnings

* reducing warnings

+++ firmware/controllers/math/engine_math.cpp
@@ -116,7 +116,7 @@ void prepareOutputSignals() {
 
 #if EFI_UNIT_TEST
 	if (verboseMode) {
-		printf("prepareOutputSignals %d %s\r\n", engineConfiguration->trigger.type, getIgnition_mode_e(engineConfiguration->ignitionMode));
+		printf("prepareOutputSignals %d %s\r\n", (int)engineConfiguration->trigger.type, getIgnition_mode_e(engineConfiguration->ignitionMode));
 	}
 #endif /* EFI_UNIT_TEST */
